refactor: Router 클래스 추가 및 라우팅 로직 분리
- Router 클래스 추출로 라우팅 관련 로직 모듈화
- 생성자에 라우트 정보를 전달
- init 함수로 초기 렌더링 및 라우팅 이벤트 등록
